requirements.txt
# ─── CAMEL AI Framework ───────────────────────────────────────────────────────
camel-ai>=0.2.0

# ─── Market Data ──────────────────────────────────────────────────────────────
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.26.0

# ─── Technical Analysis ───────────────────────────────────────────────────────
ta>=0.11.0

# ─── Reinforcement Learning ───────────────────────────────────────────────────
stable-baselines3>=2.3.0
gymnasium>=0.29.0
torch>=2.0.0

# ─── Visualization & Dashboard ────────────────────────────────────────────────
plotly>=5.18.0
matplotlib>=3.8.0

# ─── Storage & Utilities ──────────────────────────────────────────────────────
pyarrow>=14.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
